拉取最新代码

+++ phoenix-dev/src/main/java/com/dianping/maven/plugin/tools/wms/WorkspaceManagementServiceImpl.java
@@ -24,15 +24,16 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.commons.io.FileUtils;
-import org.unidal.lookup.ContainerHolder;
+import org.codehaus.plexus.DefaultPlexusContainer;
+import org.codehaus.plexus.PlexusContainer;
 import org.unidal.lookup.annotation.Inject;
 
-import com.dianping.maven.plugin.tools.misc.file.ContainerBizServerGenerator;
-import com.dianping.maven.plugin.tools.misc.file.ContainerPomXMLGenerator;
-import com.dianping.maven.plugin.tools.misc.file.ContainerWebXMLGenerator;
-import com.dianping.maven.plugin.tools.misc.file.WorkspaceEclipseBatGenerator;
-import com.dianping.maven.plugin.tools.misc.file.WorkspaceEclipseSHGenerator;
-import com.dianping.maven.plugin.tools.misc.file.WorkspacePomXMLGenerator;
+import com.dianping.maven.plugin.tools.generator.dynamic.ContainerBizServerGenerator;
+import com.dianping.maven.plugin.tools.generator.dynamic.ContainerPomXMLGenerator;
+import com.dianping.maven.plugin.tools.generator.dynamic.ContainerWebXMLGenerator;
+import com.dianping.maven.plugin.tools.generator.dynamic.WorkspaceEclipseBatGenerator;
+import com.dianping.maven.plugin.tools.generator.dynamic.WorkspaceEclipseSHGenerator;
+import com.dianping.maven.plugin.tools.generator.dynamic.WorkspacePomXMLGenerator;
 import com.dianping.maven.plugin.tools.vcs.CodeRetrieveConfig;
 import com.dianping.maven.plugin.tools.vcs.CodeRetrieverManager;
 import com.dianping.maven.plugin.tools.vcs.GitCodeRetrieveConfig;
@@ -45,10 +46,11 @@ import com.dianping.maven.plugin.tools.vcs.SVNCodeRetrieveConfig;
  */
 public class WorkspaceManagementServiceImpl implements WorkspaceManagementService {
 
-    private final static String LINE_SEPARATOR   = System.getProperty("line.separator");
-    private final static String CONTAINER_FOLDER = "phoenix-container";
-    private RepositoryManager   repositoryManager;
-   
+    private final static String  LINE_SEPARATOR   = System.getProperty("line.separator");
+    private final static String  CONTAINER_FOLDER = "phoenix-container";
+
+    @Inject
+    private RepositoryManager    repositoryManager;
     @Inject
     private CodeRetrieverManager codeRetrieverManager;
 
@@ -62,21 +64,23 @@ public class WorkspaceManagementServiceImpl implements WorkspaceManagementServic
 
             printContent("Generating phoenix workspace...", out);
 
-            if (context.getBaseDir().exists()) {
+            if (context.getBaseDir().exists() && context.isCleanFolder()) {
                 try {
                     FileUtils.cleanDirectory(context.getBaseDir());
                 } catch (IOException e) {
                     throw new WorkspaceManagementException(e);
                 }
+                printContent(String.format("Workspace folder(%s) cleared...", context.getBaseDir()), out);
             }
-            printContent(String.format("Workspace folder(%s) cleared...", context.getBaseDir()), out);
 
-            try {
-                FileUtils.forceMkdir(context.getBaseDir());
-            } catch (IOException e) {
-                throw new WorkspaceManagementException(e);
+            if (!context.getBaseDir().exists()) {
+                try {
+                    FileUtils.forceMkdir(context.getBaseDir());
+                } catch (IOException e) {
+                    throw new WorkspaceManagementException(e);
+                }
+                printContent(String.format("Workspace folder(%s) created...", context.getBaseDir()), out);
             }
-            printContent(String.format("Workspace folder(%s) created...", context.getBaseDir()), out);
 
             for (String project : context.getProjects()) {
                 Repository repository = repositoryManager.find(project);
@@ -97,7 +101,7 @@ public class WorkspaceManagementServiceImpl implements WorkspaceManagementServic
 
             printContent("Generating phoenix-container...", out);
 
-            generateContainerProject(context);
+            generateContainerProject(context, out);
 
             printContent("Generating phoenix-workspace pom...", out);
 
@@ -109,13 +113,17 @@ public class WorkspaceManagementServiceImpl implements WorkspaceManagementServic
 
             printContent("Generating ws folder...", out);
 
-            try {
-                FileUtils.forceMkdir(new File(context.getBaseDir(), "ws"));
-            } catch (IOException e) {
-                throw new WorkspaceManagementException(e);
-            }
+            File workspaceFolder = new File(context.getBaseDir(), "ws");
 
-            printContent("Phoenix workspace generated...", out);
+            if (!workspaceFolder.exists()) {
+                try {
+                    FileUtils.forceMkdir(workspaceFolder);
+                } catch (IOException e) {
+                    throw new WorkspaceManagementException(e);
+                }
+
+                printContent("Phoenix workspace generated...", out);
+            }
 
             return new File(context.getBaseDir(), CONTAINER_FOLDER);
 
@@ -150,7 +158,8 @@ public class WorkspaceManagementServiceImpl implements WorkspaceManagementServic
 
     }
 
-    private void generateContainerProject(WorkspaceContext context) throws WorkspaceManagementException {
+    private void generateContainerProject(WorkspaceContext context, OutputStream out)
+            throws WorkspaceManagementException {
         File projectBase = new File(context.getBaseDir(), CONTAINER_FOLDER);
         File sourceFolder = new File(projectBase, "src/main/java");
         File resourceFolder = new File(projectBase, "src/main/resources");
@@ -162,6 +171,11 @@ public class WorkspaceManagementServiceImpl implements WorkspaceManagementServic
 
             FileUtils.copyFileToDirectory(FileUtils.toFile(this.getClass().getResource("/byteman-2.1.2.jar")),
                     resourceFolder);
+
+            // clone gitconfig
+            GitCodeRetrieveConfig gitConfig = new GitCodeRetrieveConfig(context.getGitConfigRepositoryUrl(), new File(
+                    resourceFolder, "gitconf").getAbsolutePath(), out, context.getGitConfigRepositoryBranch());
+            codeRetrieverManager.getCodeRetriever(gitConfig).retrieveCode();
 
             // web.xml
             ContainerWebXMLGenerator containerWebXMLGenerator = new ContainerWebXMLGenerator();
@@ -196,18 +210,21 @@ public class WorkspaceManagementServiceImpl implements WorkspaceManagementServic
 
     private CodeRetrieveConfig toCodeRetrieveConfig(Repository repository, String path, OutputStream out) {
         if (repository instanceof SvnRepository) {
-            return new SVNCodeRetrieveConfig(repository.getRepoUrl(), path,
-                    out, ((SvnRepository) repository).getRevision());
+            return new SVNCodeRetrieveConfig(repository.getRepoUrl(), path, out,
+                    ((SvnRepository) repository).getRevision());
         } else if (repository instanceof GitRepository) {
-            return new GitCodeRetrieveConfig(repository.getRepoUrl(), path,
-                    out, ((GitRepository) repository).getBranch());
+            return new GitCodeRetrieveConfig(repository.getRepoUrl(), path, out,
+                    ((GitRepository) repository).getBranch());
         } else {
             return null;
         }
     }
 
     public static void main(String[] args) throws Exception {
+        PlexusContainer plexusContainer = new DefaultPlexusContainer();
         WorkspaceManagementServiceImpl wms = new WorkspaceManagementServiceImpl();
+        wms.setRepositoryManager(plexusContainer.lookup(RepositoryManager.class));
+        wms.codeRetrieverManager = plexusContainer.lookup(CodeRetrieverManager.class);
         wms.setRepositoryManager(new RepositoryManager() {
 
             @Override
@@ -229,6 +246,8 @@ public class WorkspaceManagementServiceImpl implements WorkspaceManagementServic
         // projects.add("user-web");
         context.setProjects(projects);
         context.setBaseDir(new File("/Users/leoleung/test"));
+        context.setGitConfigRepositoryBranch("master");
+        context.setGitConfigRepositoryUrl("http://code.dianpingoa.com/arch/phoenix-maven-config.git");
         wms.create(context, System.out);
     }
 }
